Added SBJson 5
This release allows scalar values at the top level; as recommended by RFC
7159, which obsoletes the original RFC 4627. Since it is a change in behaviour
I chose to bump the major version to 5.

To put SBJson on equal footing with the other parsers, this changes
the driver program to delay using the value until the full input has
been seen. This allows us to bail for the "garbage after input" cases.