Tweak lonely numbers parsing tests
When parsing numbers at the top level there is *no way* to
differentiate 42 on its own from 4200 truncated to just the first two
bytes. This only affects numbers as the other JSON datatypes are
either fixed length (true, false, null) or have unambigous delimiters
at both ends ([], {}, =”“=). (I personally consider allowing scalars
at the top level to be a mistake in RFC7159 for this reason.)

This change moves the lonely number parsing tests to become
implementation defined, and adds new "yes" tests with newlines after
the initial digits to make it obvious that the sequence of digits is
over and the token can be converted into a number.