Improve the test runner (1/2)
Catch FileNotFoundError instead of checking if a command can be run with `os.path.exists(commands[0])`. This will allow to write commands more naturally, without having to hardcode absolute paths or the `/usr/bin/env` trick.

Before this commit:
```
 "commands":["/usr/local/bin/lua", os.path.join…
```

After this commit:
```
 "commands":["lua", os.path.join…
```

Also, when an invalid binary is detected, don't try to run all the tests but stop at the first one.