Improve the test runner (2/2)
Add an optional `setup` command that is run once before running all the tests.

This can be useful to run a compilation step so that we don't need to commit huge binaries to the repository. For cross platform environment such as .NET this is even more important because we don't want to commit a (huge) ELF or Mach-O binary to be able to run the tests in a cross-platform way.

Update taking advantage of this new feature for *Json.NET* and *.NET System.Text.Json* runner will come after this pull request is merged. ;-)